Split up property list component
